requirements.txt
Flask==2.2.2
flogging==0.0.21
slack-sdk==3.18.1
slack-bolt==1.14.3
prometheus_client==0.14.1
requests==2.28.1
google-cloud-secret-manager
